services/ai-vetting/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pillow==10.3.0
python-multipart==0.0.9
torch==2.2.2
torchvision==0.17.2
